refactor: #123 TemporaryStorageService 패키지 이동 및 EventSerializer 이용하여 코드 가독성 개선
